Refactored (Changed type of id all models, changed the data type for all fields that should contain cash equivalents, changed tests)

+++ PersonalFinancialTracker/src/main/java/dev/personal/financial/tracker/service/goal/GoalServiceImpl.java
@@ -3,10 +3,14 @@ package dev.personal.financial.tracker.service.goal;
 import dev.personal.financial.tracker.dto.goal.GoalIn;
 import dev.personal.financial.tracker.dto.goal.GoalMapper;
 import dev.personal.financial.tracker.dto.goal.GoalOut;
+import dev.personal.financial.tracker.exception.goal.GoalNotFoundException;
 import dev.personal.financial.tracker.model.Goal;
 import dev.personal.financial.tracker.repository.goal.GoalRepository;
 
 import lombok.RequiredArgsConstructor;
+
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 
 /**
  * Реализация интерфейса {@link GoalService}.
@@ -23,14 +27,17 @@ public class GoalServiceImpl implements GoalService {
     }
 
     @Override
-    public GoalOut getGoalById(String id) {
+    public GoalOut getGoalById(int id) {
         Goal goal = goalRepository.findById(id);
         return GoalMapper.toDto(goal);
     }
 
     @Override
-    public GoalOut getGoalsByUserId(String userId) {
+    public GoalOut getGoalsByUserId(int userId) {
         Goal goal = goalRepository.findByUserId(userId);
+        if (goal == null) {
+            throw new GoalNotFoundException();
+        }
         return GoalMapper.toDto(goal);
     }
 
@@ -42,18 +49,25 @@ public class GoalServiceImpl implements GoalService {
     }
 
     @Override
-    public void deleteGoalByUserId(String userId) {
+    public void deleteGoalByUserId(int userId) {
+        Goal goal = goalRepository.findByUserId(userId);
+        if (goal == null) {
+            throw new GoalNotFoundException();
+        }
         goalRepository.deleteByUserId(userId);
     }
 
     @Override
-    public void updateSavedAmount(String goalId, double amount) {
+    public void updateSavedAmount(int goalId, BigDecimal amount) {
         goalRepository.updateSavedAmount(goalId, amount);
     }
 
     @Override
-    public double getProgress(String goalId) {
+    public double getProgress(int goalId) {
         Goal goal = goalRepository.findById(goalId);
-        return (goal.getSavedAmount() / goal.getGoalAmount()) * 100;
+        return goal.getSavedAmount()
+                .divide(goal.getGoalAmount(), 4, RoundingMode.HALF_UP)
+                .multiply(BigDecimal.valueOf(100))
+                .setScale(2, RoundingMode.HALF_UP).doubleValue();
     }
 }
